move to single _remote_model.py file
