--- src/fairseq2/recipes/lm/_online_finetune/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-from fairseq2.recipes.lm._online_finetune._remote_model import (
-    RayActorConfig as RayActorConfig,
-)
